+++ sdk/confidentialledger/Azure.Security.ConfidentialLedger/tests/ConfidentialLedgerClientTests.cs
@@ -1,0 +1,58 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Net;
+using System.Threading.Tasks;
+using Azure.Core;
+using Azure.Core.TestFramework;
+using NUnit.Framework;
+
+namespace Azure.Security.ConfidentialLedger.Tests
+{
+    public class ConfidentialLedgerClientTests : ClientTestBase
+    {
+        public ConfidentialLedgerClientTests(bool isAsync) : base(isAsync) { }
+        private const string transactionId = "1234";
+
+        [Test]
+        public void FailedTransaction()
+        {
+            var client = InstrumentClient(
+                new ConfidentialLedgerClient(
+                    new("https://client.name"),
+                    new MockCredential(),
+                    new ConfidentialLedgerClientOptions
+                    {
+                        Retry = { Delay = TimeSpan.Zero, MaxRetries = 0},
+                        Transport = new MockTransport(
+                            req =>
+                            {
+                                if (req.Uri.Host == "identity.accledger.azure.com")
+                                {
+                                    var cert = new MockResponse(200);
+                                    cert.SetContent(
+                                        @" {
+                                    ""ledgerTlsCertificate"": ""-----BEGIN CERTIFICATE-----\nMIIBejCCASGgAwIBAgIRANPpW17pcDYr1KnqsJH5yC8wCgYIKoZIzj0EAwIwFjEU\nMBIGA1UEAwwLQ0NGIE5ldHdvcmswHhcNMjEwMzExMDAwMDAwWhcNMjMwNjExMjM1\nOTU5WjAWMRQwEgYDVQQDDAtDQ0YgTmV0d29yazBZMBMGByqGSM49AgEGCCqGSM49\nAwEHA0IABOCPGnfcmfm5Vyax3bvg5Xqg6RUZtda0U5qpmxqGgLfL3LYJd3heTPd\u002B\n51o29pMtKJGG4cWeZ3\u002BYbhZzHnetf8WjUDBOMAwGA1UdEwQFMAMBAf8wHQYDVR0O\nBBYEFFxq\u002BImyEVh4u4BfynwnEAsbvRJBMB8GA1UdIwQYMBaAFFxq\u002BImyEVh4u4Bf\nynwnEAsbvRJBMAoGCCqGSM49BAMCA0cAMEQCIC597R3C89/IzfqjkO31XKy4Rnfy\nXauWszBChtH1v2CoAiAS0tmFNjD3fweHH8O2ySXK/tPCBTq877pIjFGwvuj2uw==\n-----END CERTIFICATE-----\n\u0000"",
+                                    ""ledgerId"": ""chrissconfidentialledger""}");
+                                    return cert;
+                                }
+                                if (req.Uri.Path.Contains($"transactions/{transactionId}/status"))
+                                {
+                                    var success = new MockResponse(500);
+                                    success.SetContent("success");
+                                    return success;
+                                }
+                                var failed = new MockResponse(200);
+                                failed.AddHeader("x-ms-ccf-transaction-id", transactionId);
+                                failed.SetContent("failed");
+                                return failed;
+                            })
+                    }));
+
+            var ex = Assert.ThrowsAsync<InvalidOperationException>(
+                async () => await client.PostLedgerEntryAsync(RequestContent.Create(new { contents = "test" }), null, true, default));
+            Assert.That(ex.Message, Does.Contain(transactionId));
+        }
+    }
+}