+++ sdk/eventhub/Azure.Messaging.EventHubs/perf/Infrastructure/EventHubsOptions.cs
@@ -6,12 +6,24 @@ using CommandLine;
 
 namespace Azure.Messaging.EventHubs.Perf
 {
+    /// <summary>
+    ///   The set of command-line options valid for Event Hubs performance tests.
+    /// </summary>
+    ///
     public class EventHubsOptions : PerfOptions
     {
-        [Option("body-size", HelpText = "Size of message body (in bytes)")]
+        /// <summary>
+        ///   The size of the event body, in bytes.
+        /// </summary>
+        ///
+        [Option("body-size", HelpText = "Size of event body (in bytes)")]
         public int BodySize { get; set; }
 
-        [Option("batch-size", HelpText = "Messages per batch")]
+        /// <summary>
+        ///   The number of events that should be included in each batch.
+        /// </summary>
+        ///
+        [Option("batch-size", HelpText = "Events per batch")]
         public int BatchSize { get; set; }
     }
 }